Add meta data, sample rate, altitude correction
Add meta data, sample rate, altitude correction
